requirements.txt
Django==5.0.1
djangorestframework==3.15.0
drf-spectacular==0.27.2
pytest==8.3.3
pytest-django==4.8.0
pytest-mock==3.14.0